fix(security): harden HMAC webhook auth with int timestamps and replay dedup
Phase 1: Replace float timestamps with int across both HMAC systems.
Replace abs() future-allowing checks with 0 <= (now - ts) <= window
to reject future timestamps (preplay prevention).

Phase 2: System 2 webhook hardening — add signature format validation
(64-char hex pre-filter), per-process replay dedup via cache.add(),
and logger.exception() for full tracebacks.

Phase 3: Add startup validation for PLATFORM_TO_PORTAL_WEBHOOK_SECRET
in both platform and portal prod.py settings.

Phase 4: Add 8 new webhook hardening tests (body tampering, future
timestamps, empty secret, missing timestamp, replay rejection, unicode
body, empty body, invalid sig format) and 12 integration tests verifying
structural properties of the hardened code.

Phase 5: Document dual HMAC architecture in AUTHENTICATION.md, update
DEPLOYMENT.md env var tables, SECURITY_CONFIGURATION.md, ARCHITECTURE.md,
data-flow diagram, and .env.example files.

+++ services/platform/tests/api/test_hmac_middleware.py
@@ -125,10 +125,10 @@ class PortalHMACTests(TestCase):
         raw_path = '/api/test/'
         portal_id = 'portal-rl'
         nonce_base = 'nonce-rl-'
-        ts = str(time.time())
+        ts = str(int(time.time()))
 
         def make_req(i: int):
-            body = json.dumps({'user_id': 1, 'customer_id': 2, 'timestamp': float(ts)}).encode()
+            body = json.dumps({'user_id': 1, 'customer_id': 2, 'timestamp': int(ts)}).encode()
             signature = self._sign(method, raw_path, body, portal_id, f'{nonce_base}{i}', ts)
             req = self.factory.post(raw_path, data=body, content_type='application/json')
             req.META['HTTP_X_PORTAL_ID'] = portal_id
@@ -153,8 +153,8 @@ class PortalHMACTests(TestCase):
         raw_path = '/api/test/'
         portal_id = 'portal-replay'
         nonce = 'replay-n'
-        ts = str(time.time())
-        body = json.dumps({'user_id': 1, 'customer_id': 2, 'timestamp': float(ts)}).encode()
+        ts = str(int(time.time()))
+        body = json.dumps({'user_id': 1, 'customer_id': 2, 'timestamp': int(ts)}).encode()
         signature = self._sign(method, raw_path, body, portal_id, nonce, ts)
 
         request1 = self.factory.post(raw_path, data=body, content_type='application/json')
